Report when the backend didn't receive a request.
Wait half a second for a backend to get a request, otherwise error. This
prevents the test from blocking forever waiting for an event that will
never occur. For instance, when the proxy handles a request itself and
does not route to a backend, while the test expects the backend to be
hit.